move view_notifications to dashboard.py  It is more logically consistent for this function to reside in dashboard.py than in jobs.py
